001/requirements.txt
azure-storage-blob==12.19.0
azure-identity==1.15.0